+++ src/app/admin/admin.module.ts
@@ -5,10 +5,12 @@ import { AdminRoutingModule } from './admin-routing.module';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '../shared/shared.module';
 import { FormsModule } from '@angular/forms';
+import { ManageProductsListComponent } from './components/manage-products-list/manage-products-list.component';
+import { ManageProductsItemComponent } from './components/manage-products-item/manage-products-item.component';
 
 
 @NgModule({
-  declarations: [AdminRoutingModule.components],
+  declarations: [AdminRoutingModule.components, ManageProductsListComponent, ManageProductsItemComponent],
   imports: [
     CommonModule,
     SharedModule,