+++ src/app/admin/components/index.ts
@@ -2,3 +2,4 @@ export * from './admin/admin.component';
 export * from './manage-orders/manage-orders.component';
 export * from './manage-products/manage-products.component';
 export * from './manage-products-form/manage-products-form.component';
+export * from './manage-products-list/manage-products-list.component';