+++ src/app/admin/components/manage-products-form/manage-products-form.component.ts
@@ -1,22 +1,23 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductModel } from 'src/app/products/models/product.model';
-import { ProductCategory } from 'src/app/products/models/product-category.enum';
-import { LocalStorageService } from 'src/app/core/services/local-storage.service';
 import { Router, ActivatedRoute } from '@angular/router';
-import { ProductsService } from 'src/app/products/services/products.service';
 import { pluck } from 'rxjs/operators';
+import { ProductsPromiseService, ProductsObservableService } from 'src/app/products';
 
 @Component({
   selector: 'app-manage-products-form',
   templateUrl: './manage-products-form.component.html',
   styleUrls: ['./manage-products-form.component.css']
 })
+
 export class ManageProductsFormComponent implements OnInit {
   product: ProductModel;
   originalProduct: ProductModel;
-  constructor(public productsService: ProductsService,
-              private route: ActivatedRoute,
-              private router: Router) { }
+  constructor(
+    public productsPromiseService: ProductsPromiseService,
+    public productObservableService: ProductsObservableService,
+    private route: ActivatedRoute,
+    private router: Router) { }
 
   ngOnInit() {
     this.route.data.pipe(pluck('product')).subscribe((product: ProductModel) => {
@@ -25,16 +26,15 @@ export class ManageProductsFormComponent implements OnInit {
   }
 
   onSaveProduct() {
-      const product = { ...this.product };
-      if (product.id === 0) {
-      this.productsService.addProduct(this.product);
-      this.onGoBack();
+    const product = { ...this.product };
+    if (product.id === 0) {
+      this.productsPromiseService.addProduct(this.product)
+        .then(c => this.onGoBack());
     } else {
-      this.productsService.editProduct(this.product);
-      this.onGoBack();
+      this.productObservableService.updateProduct(this.product).subscribe(c => this.onGoBack());
     }
   }
   onGoBack() {
-    this.router.navigate(['./../'], { relativeTo: this.route});
+    this.router.navigate(['admin', 'products']);
   }
 }