+++ src/app/layout/components/about/about.component.ts
@@ -6,6 +6,7 @@ import { ConfigOptionsModel } from 'src/app/core/models/config-options.model';
 import { ConstantsService, constInstance } from 'src/app/core/services/constants.service';
 import { genData, GeneratorFactoryService } from 'src/app/core/services/generator-factory';
 import { GeneratorService } from 'src/app/core/services/generator.service';
+import { AppSettingsService } from 'src/app/core/services/app-settings.service';
 
 @Component({
   selector: 'app-about',
@@ -24,13 +25,25 @@ export class AboutComponent implements OnInit, OnDestroy {
   constructor(
     private storage: Storage,
     private constantService: ConstantsService,
+    private localStorageService: LocalStorageService,
+    public appSettingsService: AppSettingsService,
+
     @Inject(genData) private factoryData: string,
     @Optional() private configOptionsService: ConfigOptionsService
   ) {
   }
 
   ngOnInit() {
-    this.applicationInfo = this.constantService.getAllData();
+    this.applicationInfo = JSON.parse(this.appSettingsService.getSettingsFromLS());
+    if (!this.applicationInfo) {
+      this.appSettingsService.getSettingsFromFile().subscribe(
+        data => {
+          this.applicationInfo = data ? data : this.constantService.getAllData();
+          if (data) {
+            this.localStorageService.setItem('appSettings', JSON.stringify(data));
+          }
+        });
+    }
 
     this.storage.setItem('info', JSON.stringify(this.applicationInfo));